merge powerarrow-dark theme as branch of powerarrow
These two themes appear to be closely related and share quite a bit of
code. Maintaining them as two separate projects seems like a hassle.
Merging the two into one project with an optional light/dark selector
seems ideal, but at least having it as a branch in the same git repo is
a step in that direction. Comparing the differences in the code base
will be much easier and bugfixes, refinements and documentation will be
much easier to port both directions.